Update to site logo
background-clip: text is awesome.  Too bad the effect is impossible to
emulate in other browsers without resorting to images.  Firefox was
wickedly slow to render SVG text with a image fill and Opera made it
impossible to line up text shadows.

An image it is.  But it'll be a good image.

Unfortunately the site's background color needs to be baked into the
logo image to make this method work.