Removes save method's tag purity check
This check is almost always working with insufficient tag data, requiring that create and update views do a secondary crunch anyway.  Worse, if a user removes a tag from a wallpaper and saves the change, the tag assessment will still include the removed tag.

Better to just remove it and require that wallpaper create and update views call `assess_tag_purity()` after saving the object.