Adds ActiveProfileMiddleware
There are as many as three places where the active profile was queried
on every request.  This middleware now assigns the active profile to
``user.active_profile`` to be accessed on demand.

The active profile is able to exist for anonymous users, since the
"default" profile is an unsaved instance that is generated at runtime
with default values on its attributes.